Refactored tests for runners into the runners folder
