+++ src/__tests__/environment.ts
@@ -18,15 +18,15 @@ test('Function params and body identifiers are in different environment', () => 
   const context = mockContext(Chapter.SOURCE_4)
   context.prelude = null // hide the unneeded prelude
   const parsed = parse(code, context)
-  const it = evaluate(parsed as any as Program, context, {
-    envSteps: -1,
+  evaluate(parsed as any as Program, context, {
+    envSteps: 13,
     stepLimit: 1000,
     isPrelude: false
   })
-  const stepsToComment = 13 // manually counted magic number
-  for (let i = 0; i < stepsToComment; i += 1) {
-    it.next()
-  }
+  // const stepsToComment = 13 // manually counted magic number
+  // for (let i = 0; i < stepsToComment; i += 1) {
+  //   it.next()
+  // }
   context.runtime.environments.forEach(environment => {
     expect(environment).toMatchSnapshot()
   })
