+++ src/runner/__tests__/files.ts
@@ -59,7 +59,7 @@ describe('runFilesInContext', () => {
 
   it('returns ModuleNotFoundError if entrypoint file does not exist', async () => {
     const files: Record<string, string> = {}
-    await await runFilesInContext(files, '/a.js', context)
+    await runFilesInContext(files, '/a.js', context)
     expect(parseError(context.errors)).toMatchInlineSnapshot(`"Module '/a.js' not found."`)
   })
