+++ src/validator/__tests__/validator.ts
@@ -5,7 +5,8 @@ import { parse } from '../../parser/parser'
 import { Chapter, type NodeWithInferredType } from '../../types'
 import { getVariableDeclarationName } from '../../utils/ast/astCreator'
 import { stripIndent } from '../../utils/formatters'
-import { astTester, expectTrue } from '../../utils/testing'
+import { expectTrue } from '../../utils/testing'
+import { astTester } from '../../utils/testing/testers'
 import { simple } from '../../utils/walkers'
 import { checkForUndefinedVariables, validateAndAnnotate } from '../validator'
 import { UndefinedVariable } from '../../errors/errors'
